requirements.txt
debugpy
pylint-odoo==7.0.0
pylint-gitlab
pylint-exit
debugpy
openpyxl